Refactored and broke Class Scraper into multiple files
