+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassData/GetActivity.ts
@@ -1,0 +1,15 @@
+/**
+ * Extracts the activity field, makes sure that it exists
+ * @param { string } data:  Line which contains the activity field
+ * @returns { string }: Type of activity the class is
+ */
+const getActivity = (data: string): string => {
+  const activity = data
+  if (!activity) {
+    throw new Error('Unknown activity: ' + activity)
+  }
+
+  return activity
+}
+
+export { getActivity }
