+++ scraper/src/PageScraper/ChunkScraper/ClassScraper/ClassData/GetMode.ts
@@ -1,0 +1,14 @@
+/**
+ * Extracts the mode of class delivery
+ * @param { string } mode: Line which contains the mode field
+ * @returns { string }: Mode of delivery for a class
+ */
+const getMode = (mode: string): string => {
+  if (!mode || mode === ' ') {
+    throw new Error('Invalid Mode: ' + mode)
+  }
+
+  return mode
+}
+
+export { getMode }
